+++ src/Controller/User/DeleteTokenController.php
@@ -8,6 +8,7 @@ use App\Response\NoContentResponse;
 use App\Security\AuthProvider;
 use App\Service\ElementManager;
 use LogicException;
+use Predis\Client;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -17,6 +18,7 @@ class DeleteTokenController extends AbstractController
     public function __construct(
         private ElementManager $elementManager,
         private AuthProvider $authProvider,
+        private Client $redisClient,
         private Client401UnauthorizedExceptionFactory $client401UnauthorizedExceptionFactory,
         private Client404NotFoundExceptionFactory $client404NotFoundExceptionFactory
     ) {
@@ -39,19 +41,27 @@ class DeleteTokenController extends AbstractController
             throw $this->client401UnauthorizedExceptionFactory->createFromTemplate();
         }
 
+        $hashedToken = $this->authProvider->getHashedToken();
+        if (null === $hashedToken) {
+            throw $this->client401UnauthorizedExceptionFactory->createFromTemplate();
+        }
+
         $tokenUuid = $this->authProvider->getTokenUuid();
         if (null === $tokenUuid) {
             throw new LogicException('Token must be provided.');
         }
 
-        $element = $this->elementManager->getElement($tokenUuid);
-        if (null === $element) {
+        $tokenElement = $this->elementManager->getElement($tokenUuid);
+        if (null === $tokenElement) {
             throw $this->client404NotFoundExceptionFactory->createFromTemplate();
         }
-        $this->elementManager->delete($element);
+        $this->elementManager->delete($tokenElement);
         $this->elementManager->flush();
 
-        // todo: remove cached token from redis
+        $this->redisClient->expire(
+            $this->authProvider->getRedisTokenKeyFromHashedToken(
+                $hashedToken
+            ), 0);
 
         return new NoContentResponse();
     }